refactor: 모아보기 공통 Dto로 이름 변경
