+++ workit/src/main/java/workit/dto/project/AllProjectCollectionDetailResponseDto.java
@@ -1,15 +1,16 @@
 package workit.dto.project;
 
 import lombok.Data;
+import workit.dto.collection.CollectionDetailResponseDto;
 
 import java.util.List;
 
 @Data
 public class AllProjectCollectionDetailResponseDto {
     private String projectTitle;
-    private List<ProjectCollectionDetailResponseDto> works;
+    private List<CollectionDetailResponseDto> works;
 
-    public AllProjectCollectionDetailResponseDto(String projectTitle, List<ProjectCollectionDetailResponseDto> works) {
+    public AllProjectCollectionDetailResponseDto(String projectTitle, List<CollectionDetailResponseDto> works) {
         this.projectTitle = projectTitle;
         this.works = works;
     }
